Add constraint engine and an example.
See: https://developers.google.com/optimization

Why: Adding a constraint engine to do musical constraint programming.
Since constraint programming can be resource intensive, I am relying on
Google's OR-Tools (Operations Research Tools) which has a built-in
constraint solver. The OR-Tools API is a little too broad for musical
applications (at least my own applications!) so I am creating a facade
to make it easier. Additionally, I am providing a few helper functions
and making instantiating a solver much simpler.

Additionally I added an example of how to use the constraint library in
the `examples` directoy. This example is from a piece that I am
currently working on, that I is constrainted to palindromic 12-tone
chords (in fixed pitch space).

How: Adding a new file in the allegro module called constraints.py. This
file has a class `Problem` that provides an API for modelling and
solving constraint satisfaction problems. Google's OR-Tools is very
broad, but I provided access to the `model` object in to keep things
flexible. I added tests.

Tags: constraints, or-tools