Add a caching module.
Why: I need a way to save the expensive results from the constraint
programming library to disk and load them back up again.

Also, add hashing algorithm to constraint problem. I'm thinking of doing
an "autocache", but it'll probably be too big of a headache.

How: Adding a caching module. This creates a cache directory in
XDG_CACHE_HOME or if that doesn't exist, $HOME/.cache. It adds a clas
`Cache` which has two methods, `save` and `load`. `save` writes any
JSON-serializable to to the cache directory. `load` looks for a json
file matching the name parameter and deserializes the JSON into python.

Tags: caching, JSON, XDG