Update the katabasis constraint programming example.
Why: Katabasis is a piece I am currently working on. This example uses
the constraint engine to generate fixed-pitch rows that exhibit cetain
qualities. Additionally it adds the music21 project for it's analysis
qualities. Lastly, it fixes a few bugs in constraint.py and provides a
new interface for playing "chords" in the midi module.

How: Flesh out the katabasis example in examples/katabasis.py.

Tags: examples, midi, constraint