Update various dependencies
Why: Trying to create a reproducible dev environment using Nix shell and
poetry.

How: I updated the shell.nix file to use a direct tarball instead of a
nix channel. It appears that nix-channels are going out of favor and
they have lots of bugs and reproducibility issues. So, I switched to the
tarball and updated to python 3.11. I added back ortools and rtmidi as
well.

Tags: dev, environment, nix